requirements.txt
matplotlib
cython==0.28.2
kivy
pillow